YTI-468 , Bugfix , continuation part.
It occurred in testing that only relatively old members worked correctly
if their values were set to empty in the UI. If a new member was created
and then updated to have empty values, the update still failed.

This is a side-effect of larger refactorings done in the fall of 2018
but it was fixed now so that the value is never null in the front, it is
always an empty string with length of zero.