YTI-922 , bugfix , start using updated version of common-ui.
The change in common-ui is that in an interim component a placement
input value now also has a default value. It was not enough that there
was a default value in the Ngb-component itself, any missing placement
value down the stream of inputs results in that default value getting
overwritten with an empty value.