Angular 12 upgrade
This commit upgrades the project to angular 12. Upgraded yti-common-ui
is a prerequisite.

The final commit is squashed from several separate commits to make it
clear when a change originates from this upgrade.

Updated dependencies:

* Primary upgrade has been Angular to version 12.
* Another necessary upgrade was ng-bootstrap to version 10.0.0, and
  Bootstrap to 4.5.0
* Other library versions updated to reasonably recent versions

lazyFor:

[lazyFor](https://github.com/robianmcd/angular-lazy-for) is an older
library, which has not been maintained with newer angular versions. The
library still works, but needs to be built for new angular. Since the
package in the npm registry is not up to date, the directive has been
added directly to the code.

This may still have to change in near future, possibly by replacing
lazyFor with some other solution.

Bootstrap TabSets:

Tabset has been replaced with Nav directives. This caused minimal
changes, although in some cases (code-scheme-extensions) styles of
sub-tabs needed some tweaking to render correctly.

code-scheme-extensions also had some issue with infinite loop, which
seems to be fixed by changing function to a member variable, when it's
read from a template.

require():

Usages of require() have been replaced with import or with other means.

yti-common-ui dependency:

The yti-common-ui was previously included as-is, but is now converted to
an Angular Library project. This was required to make things work with
Ivy.

This means the yti-common-ui needs to be be built beforehand with "npm
run build", and possibly published as an npm package.

To make this possible, yti-common-ui is published as an npm package in
the github npm repository. Developers can also manually point
packages.json to the dist/yti-common-ui/ path. Dockerfile also does
this automatically.

The npm package source for the @vrk-yti scope is configured in .npmrc
and .yarnrc, so it should never attempt to load from public registry.

https://docs.github.com/en/packages/working-with-a-github-packages-registry/working-with-the-npm-registry

Other:

Some other small tweaks not specifically mentioned here, due to typescript syntax, change of bootstrap styles, etc.

YTI-1263