fix programmatically changing tabs
Tabs using ngbNav were setup incorrectly, since the "ngbNabItem"
attribute was using data binding, while the value given to it was not a
variable.

As a result, tabs were internally give auto-generated names, which meant
selecting the tabs was only possible using those generated names.

Possible fix could've been to provide a string to it:

  [ngbNavItem]="'foo'"

But since the data binding is not really useful without a variable, a
simpler format is enough:

  ngbNavItem="foo"

In particular, this commit fixes an issue with automatically selecting
"extension_members_tab" when "goToMembersTab" query parameter is
provided to extension-component.

YTI-1566