Fix "no active tab" problem in code scheme
In general, one tab should be active all the time in all kind of
tab views. There are exceptions but this is the rule of thumb.

This time it was possible to end up in a situation where active
tab === undefined. It was caused sometimes in a situation where
active tab was not requested in query parameters.

Fixed it by selecting previously selected tab instead of undefined.
If previously selected tab is undefined too, it is still possible
to end up in the same situation where active tab === undefined.

YTI-1576